Remove formatDate method from StoriesAdapter class to FormatUtils class
