+++ app/src/main/java/mitya/yahnc/FormatUtils.java
@@ -1,0 +1,30 @@
+package mitya.yahnc;
+
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
+import java.util.Date;
+
+/**
+ * Created by Mitya on 16.07.2016.
+ */
+public class FormatUtils {
+    public static String formatDate(long time) {
+        Date dateTime = new Date();
+        dateTime.setTime(time * 1000);
+        Calendar calendar = Calendar.getInstance();
+        calendar.setTime(dateTime);
+        Calendar today = Calendar.getInstance();
+        Calendar yesterday = Calendar.getInstance();
+        yesterday.add(Calendar.DATE, -1);
+        DateFormat timeFormatter = new SimpleDateFormat("hh:mma");
+
+        if (calendar.get(Calendar.YEAR) == today.get(Calendar.YEAR) && calendar.get(Calendar.DAY_OF_YEAR) == today.get(Calendar.DAY_OF_YEAR)) {
+            return "Today " + timeFormatter.format(dateTime);
+        } else if (calendar.get(Calendar.YEAR) == yesterday.get(Calendar.YEAR) && calendar.get(Calendar.DAY_OF_YEAR) == yesterday.get(Calendar.DAY_OF_YEAR)) {
+            return "Yesterday " + timeFormatter.format(dateTime);
+        } else {
+            return dateTime.toString();
+        }
+    }
+}
